Pseudo-multiserver support. "Claim" feature.
The bot now writes the access and owner files to their respective
directories determined by the config file, so that you don't have to
keep the binary multiple times in seperate directories for seperate
servers. Now you can have just one, and the bot will load it accordingly
to the config file. Also added "claim" for when the bot first joins the
server, you can specify a password in the config file that automatically
adds you as the highest-ranking user.

Also fixed a minor error in the settings parser that would cause the
last line to be corrupt.